requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
psycopg2-binary==2.9.10
python-dotenv==1.1.1
regex==2025.9.18
requests==2.32.5
tiktoken==0.11.0
urllib3==2.5.0
